fix: address PR review
