+++ back_end/src/index.ts
@@ -4,8 +4,6 @@ import express, { type Request, type Response } from "express";
 import cors from "cors";
 import { env } from "./config/env";
 import cookieParser from "cookie-parser";
-import MyProfile from "./routes/user.route";
-import changeRole from "./routes/user.route";
 import userRoutes from "./routes/user.route";
 
 const app = express();
@@ -15,15 +13,13 @@ app.use(cookieParser());
 app.use(
     cors({
         origin: env.ALLOWED_ORIGIN,
-        methods: ["GET", "POST", "PATCH","DELETE"],
+        methods: ["GET", "POST", "PATCH", "DELETE"],
         credentials: true,
     }),
 );
 
 app.use("/auth", authRoutes);
 app.use("/team", Team);
-app.use("/me", MyProfile);
-app.use("/role", changeRole);
 app.use("/user", userRoutes);
 
 app.get("/", (_req: Request, res: Response) => {
@@ -33,5 +29,5 @@ app.get("/", (_req: Request, res: Response) => {
 app.listen(env.PORT, () => {
     console.log(`Server is running on port ${env.PORT}`);
 });
-    
+
