+++ back_end/src/routes/team.route.ts
@@ -2,9 +2,7 @@ import { Router,Request, type Response } from "express";
 import { authMiddleware } from "../middleware/auth-middleware";
 import { z } from "zod";
 import { role } from "../types/user.type";
-import { createTeamService } from "../services/teamService";
-import { deleteTeamById } from "../services/teamService";
-import { addMemberService } from "src/services/addMemberService";
+import { createTeamService, deleteTeamById, addMemberService } from "../services/teamService";
 
 const router = Router();
